refactor(web): move Toggle to design system
