+++ web/src/components/ui/toggle-group.tsx
@@ -5,7 +5,7 @@ import * as React from "react";
 import * as ToggleGroupPrimitive from "@radix-ui/react-toggle-group";
 import { type VariantProps } from "class-variance-authority";
 import { cn } from "@/src/utils/tailwind";
-import { toggleVariants } from "@/src/components/ui/toggle";
+import { toggleVariants } from "@/src/components/design-system/Toggle/Toggle";
 
 const ToggleGroupContext = React.createContext<
   VariantProps<typeof toggleVariants>
