+++ web/src/features/datasets/components/DatasetRunAggregateColumnHelpers.tsx
@@ -9,7 +9,7 @@ import { type Row } from "@tanstack/react-table";
 import React, { useEffect, useRef, useState } from "react";
 import { useDebounce } from "@/src/hooks/useDebounce";
 import { type ScoreColumn } from "@/src/features/scores/types";
-import { Toggle } from "@/src/components/ui/toggle";
+import { Toggle } from "@/src/components/design-system/Toggle/Toggle";
 import { useRouter } from "next/router";
 import { cn } from "@/src/utils/tailwind";
